Refactor engagement actions: consolidate components and remove deprecated classes
